fix: 시세 안정성 평가 연결 누락 복구 #302

+++ backend/src/main/java/org/tajiro/favorite/dto/FavoriteDTO.java
@@ -17,7 +17,7 @@ public class FavoriteDTO {
     private String tradeType;
     private Integer deposit;
     private Integer monthlyRent;
-    private Integer evaluationScore;
+    private BigDecimal evaluationScore;
     private String imageUrl;
     private String propertyType;
     private BigDecimal areaM2;
